functions/parselmouth/requirements.txt
praat-parselmouth==0.4.0
ffmpeg-python==0.2.0
python-magic        0.4.24